Implement WordClass - Mod compatibility checks
Compatibility checks are required to ensure the correct word
processing by Generator.Transform method. Some words are members
of multiple main lists. This may lead to incorrect behavior.

For example, up to this change, the verb 'produce' combined with
MOD_PLURAL and MOD_PRESENT_SIMPLE would not be transformed,
because 'produce' is also an uncountable noun. The code did not
discriminate parts of speech and this transformable verb would be
treated as a noun.