Recompile the list of non-comparable adjectives using a more up-to-date source
The old list was built using the page titles in 'Category:Non-comparable_adjectives'
from Simple English Wiktionary, which was last updated in 2006.

The new list is much larger and more accurate. It is comprised of the page titles
in 'Category:English_uncomparable_adjectives' from the regular English Wiktionary,
last updated in 2022.