Add MOD_INDEF_SILENT to support MOD_INDEF in phrases
Indefinite article is not always placed before a noun. It can be
combined with an adjective that describes a noun. In such case,
the noun must be compatible with an indefinite article.

MOD_INDEF_SILENT serves as a constraint for noun generation
process. It ensures that groups of nouns incompatible with
MOD_INDEF - uncountable and plural-only nouns - are ignored
by the generator.